Z_IsEnoughFreeMemory() int32_t -> uint16_t

+++ z_zone.h
@@ -31,7 +31,7 @@
 
 void Z_Init(void);
 void Z_Shutdown(void);
-boolean Z_IsEnoughFreeMemory(int32_t size);
+boolean Z_IsEnoughFreeMemory(uint16_t size);
 void __far* Z_TryMallocStatic(int32_t size);
 void __far* Z_MallocStatic(int32_t size);
 void __far* Z_MallocStaticWithUser(int32_t size, void __far*__far* user);
